% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{anyUndefined}
\alias{anyUndefined}
\title{Checks if any parameter is NULL or a list of length zero}
\usage{
anyUndefined(...)
}
\arguments{
\item{...}{list of parameters to check}
}
\value{
boolean
}
\description{
When Shiny starts all reactive expression are run through with empty input parameters, 
e.g. for initialising variables for the expression. Mostly the output can only generated if 
all input parameters are set/given. Thus \code{anyUndefined} delivers \code{FALSE} is any 
cecked \code{input} parameter is not set.
}
\examples{
anyUndefined(list(), NULL)
}
