% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceInterval.R
\name{CI.mean}
\alias{CI.mean}
\title{CI.mean}
\usage{
CI.mean(sample, sd = NULL, level = 0.95, B = 0)
}
\arguments{
\item{sample}{numeric: data vector}

\item{sd}{numeric: if given then the standard deviation is assumed to be known (default: \code{NULL})}

\item{level}{numeric: level of the confidence interval() default: \code{0.95})}

\item{B}{numeric: if larger than zero then an bootstrapped confidence interval is computed (default: \code{0})}
}
\value{
confidence interval
}
\description{
Computes the confidence interval for the mean of the sample
}
\examples{
x <- runif(100)
CI.mean(x)
CI.mean(x, sd=1/sqrt(12))
CI.mean(x, B=1000)
}
