% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{contingencyTable}
\alias{contingencyTable}
\title{contingencyTable}
\usage{
contingencyTable(tab)
}
\arguments{
\item{tab}{a contingency table}
}
\value{
a contingency table of observed frequencies
}
\description{
Creates a "random" contingency table of size \code{ncol} times \code{nrow} which has a total number of 
\code{n} observations. For \code{n} a range can be given; if a single value is given than the range 
\code{round(n*0.8)} till \code{round(n*1.2)} is used. With \code{digits} the number of digits after the 
comma for the expected frequencies can be given.
With the following parameter(s) a target range of association can be given for the contingency table
\describe{
\item{\code{CC}}{target range for the corrected contingency coefficient}
\item{\code{V}}{target range for Cramer's V}
\item{\code{C}}{target range for the contingency coefficient}
\item{\code{chisq}}{target range for the chi square value}
}
For any parameter (except \code{chisq}) given as a single value the range is computed as parameter plus/minus 0.05.
If \code{chisq}) is given as a single value the range is computed as parameter 
plus/minus \eqn{\sqrt{(ncol-1)(nrow-1)})}.
If more than one target parameter is given then a join of the corresponding chi square ranges will be made.
}
\note{
It is not guaranteed that for all parameters combinations a solution can be found. The function stops
after one thousand iterations to find an appropriate table of expected or observed frequencies with an error! 
You may increase/enlarge the range of \code{n}, \code{cc} or increase \code{digits}.
}
\examples{
ct <- contingencyTable(HairEyeColor[,,"Female"])
}
