% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{distrFunc}
\alias{distrFunc}
\alias{distrFunc.default}
\title{distrFunc}
\usage{
distrFunc(type, ...)

\method{distrFunc}{default}(type, func = "d", discrete = F, supp = c(NA,
  NA), xlim = c(NA, NA), ylim = c(NA, NA), tol = .Machine$double.eps^0.25,
  pretty = T, ...)
}
\arguments{
\item{type}{character: distribution name}

\item{...}{further parameters depending on the distribution}

\item{func}{character: type of function (default: \code{'d'})}

\item{discrete}{logical: if distribution is discrete or not (default:
\code{FALSE})}

\item{supp}{range: the support of the distribution function (default: \code{c(NA, NA)}, detect automatically)}

\item{xlim}{range: set the range of the non-zero density/mass function
(default: \code{c(NA, NA)}, detect automatically)}

\item{ylim}{range: set the range of the y-values 
(default: \code{c(NA, NA)}, detect automatically)}

\item{tol}{numeric: tolerance value (default:
\code{.Machine$double.eps^0.25})}

\item{pretty}{logical: use the function \code{pretty} to determine good limits}
}
\value{
a distribution object
}
\description{
Creates a distrFunc object from a distribution name. The density/mass (\code{func='d'}), 
the cumulative distribution (\code{func='p'}) and the quantile distribution (\code{func='q'}) function 
are possible choices.
}
\details{
The standard distributions from the \pkg{stats} package (except the multinomial
distribution) are supported 
\tabular{ll}{ 
\code{beta}    \tab beta distribution \cr 
\code{binom}   \tab binomial (including Bernoulli) distribution \cr 
\code{cauchy}  \tab Cauchy distribution \cr 
\code{exp}     \tab exponential distribution \cr 
\code{f}       \tab F distribution \cr 
\code{gamma}   \tab gamma distribution \cr 
\code{geom}    \tab geometric distribution \cr
\code{hyper}   \tab hypergeometric distribution \cr 
\code{lnorm}   \tab log-normal distribution \cr 
\code{nbinom}  \tab negative binomial distribution \cr 
\code{norm}    \tab normal distribution \cr 
\code{pois}    \tab Poisson distribution \cr 
\code{t}       \tab Student's t distribution \cr 
\code{unif}    \tab uniform distribution \cr 
\code{weibull} \tab Weibull distribution 
}

The function computes the range of the non-zero density/mass function and the
maximum of the density/mass function.

For a standard distribution of the \pkg{stats} package you need only
  to provide the name in the parameter \code{type}. For a non-standard distribution with
  the name 'wwww' is expected that the following functions exist: 
  \tabular{ll}{
  \code{dwwww(x, ...)} \tab density/mass function \cr
  \code{pwwww(q, ...)} \tab distribution function \cr
  \code{qwwww(p, ...)} \tab quantile function
  } 
  A discrete distribution must be defined for non-negative integers.
  
  The range of non-zero density/mass function is computed by \code{qwwww(0,
  ...)} and \code{qwwww(1, ...)}. If any of these quantiles returns
  plus/minus infinity then \code{qwwww(tol, ...)} or \code{qwwww(1-tol, ...)}
  is used. Alternatively you may set the range by the parameter \code{xlim} 
  and a \code{NA} value implies the quantile function should be used.
}
\examples{
d <- distrFunc("norm")
d
}
