% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyTable.R
\name{frequencyTable}
\alias{frequencyTable}
\alias{frequencyTable.default}
\title{frequencyTable}
\usage{
frequencyTable(obj, ...)

\method{frequencyTable}{default}(obj, breaks = 5, ...)
}
\arguments{
\item{obj}{numeric: a vector of numeric values to group}

\item{...}{currently unused}

\item{breaks}{numeric: breaks for creating a frequency table}
}
\value{
\code{NULL}
}
\description{
frequencyTable
}
\examples{
ft <- frequencyTable(rnorm(100))
}
