% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateObject.R
\name{generate}
\alias{generate}
\alias{generateObject}
\alias{generateObject.default}
\title{generate, generateObject}
\usage{
generate(obj, ...)

generateObject(obj, ...)

\method{generateObject}{default}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{further to generate a new object from \code{obj}}
}
\value{
a new object or \code{NULL} is generate was not success

\code{NULL}
}
\description{
\code{generateObject} generates a object from \code{obj} with further parameters given in \code{...}. 
\code{generate} is just a shorthand for \code{generateObject}.
}
\note{
There exists no default method to generatea new object. An error will be thrown.
}
\examples{
x  <- rnorm(37)
ft <- frequencyTable(x)
generate(ft)
}
