% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{generateObject.contingencyTable}
\alias{generateObject.contingencyTable}
\title{generateObject.contingencyTable}
\usage{
\method{generateObject}{contingencyTable}(obj, n = NULL, CC = c(0.45, 0.55),
  V = NULL, C = NULL, chisq = NULL, digits = 0, verbose = F)
}
\arguments{
\item{obj}{contingency table object}

\item{n}{integer range: total number of observations}

\item{CC}{numeric range: target corrected contingency coefficient (default: c(0.45, 0.55))}

\item{V}{numeric range: target Cramer's V (default: NULL)}

\item{C}{numeric range: target contingency coefficient (default: NULL)}

\item{chisq}{numeric range: target chi square value (default: NULL)}

\item{digits}{integer: number of digits after the comma for the expected frequencies (default: 0)}

\item{verbose}{integer: verbosity of function (default: 0)}
}
\value{
a contingency table with the approximately required properties
}
\description{
Create a contingency table with certain properties.
}
\examples{
ft  <- contingencyTable(table(round(runif(10)), round(runif(10))))
ft2 <- generate(ft)
}
