% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyTable.R
\name{generateObject.frequencyTable}
\alias{generateObject.frequencyTable}
\title{generateObject.frequencyTable}
\usage{
\method{generateObject}{frequencyTable}(obj, digits = 2, n = c(0.75, 1.25) *
  length(obj$x), zero = TRUE, ...)
}
\arguments{
\item{obj}{frequency table object}

\item{digits}{numeric: number of significant digits for relative frequency (default: \code{2})}

\item{n}{range: range of observation numbers}

\item{zero}{logical: are lasses with zero observations allowed (default: \code{TRUE})}

\item{...}{currently unused}
}
\value{
a frequency table object
}
\description{
generateObject.frequencyTable
}
\examples{
ft <- frequencyTable(runif(100))
ft
ft2 <- generate(ft)
ft2
}
