% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInputs.R
\name{getInputs}
\alias{getInputs}
\title{Input names for the widget elements}
\usage{
getInputs(env)
}
\arguments{
\item{env}{widget}
}
\value{
The names of the inputs which are used inside the widget
}
\description{
Input names for the widget elements
}
\examples{
w <- widgetFontSize('widget')
getInputs(w)
}
