% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetObserve.R
\name{getShinyInfo}
\alias{getShinyInfo}
\title{getShinyInfo}
\usage{
getShinyInfo(funcname)
}
\arguments{
\item{funcname}{character: name of shiny function}
}
\value{
information about a specific shiny input function
}
\description{
Information for input functions in the package \code{shiny} are stored in the library \code{mmstat}.
To access the information this function is used.
}
\examples{
getShinyInfo('sliderInput')
}
