% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetLanguage.R
\name{getText}
\alias{getText}
\title{Text translation}
\usage{
getText(msg, env = NULL)
}
\arguments{
\item{msg}{text vector to translate}

\item{env}{language widget}
}
\value{
For all elements of the text vector a translation will be delivered.
If no translation is available then the original text will be given back.
}
\description{
Each element of the text vector \code{msg} is translated in the current target language. 
If the translation of a text element is not possible then the original text is delivered.
}
\details{
For the translations must be PO files provided. If a Shiny app 
is run locally (not under a server) \code{getText} writes in the current directory a file
\code{getText.log} which contains elements of \code{msg} in a form such that it can be 
parsed with \code{xgettext getText.log}. You may use \code{Poedit} to create and update
your PO files.
}
\note{
Since the analysis of the PO files is done by an R function in the package, 
not all features which PO file allows are available.
}
\examples{
getText('GERMAN', NULL)
}
\seealso{
\code{\link{selectInput}}
}
