% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetDataset.R
\name{getVariables}
\alias{getVariables}
\title{Returns variable(s) of a data set}
\usage{
getVariables(env, input, session)
}
\arguments{
\item{env}{data set widget}

\item{input}{input object: current values of all widgets in your app. Is the first parameter of the \code{server} function.}

\item{session}{session object: is the third optional parameter of the \code{server} function.}
}
\value{
a list of data frames, one for each var element in \code{widgetDataset}
}
\description{
Returns variable(s) of a data set
}
\seealso{
\code{\link{widgetDataset}}
}
