% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{includeWidget}
\alias{includeWidget}
\title{includeWidget}
\usage{
includeWidget(dest, src)
}
\arguments{
\item{dest}{widget to be added to}

\item{src}{widget to be added from}
}
\description{
Includes the content of the \code{src} widget into the widget \code{dest}. 
Elements which are not included are: language widget and internal data.
}
\examples{
w1 <- widgetHypergeometric('hyper')
w2 <- widgetBinomial('binomial')
includeWidget(w1, w2)
summary(w1)
}
