% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{interval}
\alias{interval}
\title{interval}
\usage{
interval(dist, left, right, shift = 0, n = 201, ...)
}
\arguments{
\item{dist}{distribution function}

\item{left}{numeric: left interval border}

\item{right}{numeric: right interval border}

\item{shift}{numeric: shift of x in case of a mass function}

\item{n}{numeric: number of points in case of a continuous distribution}

\item{...}{further named graphical parameters}
}
\description{
Plots interval(s) according to the selected distribution function. Please note that for a continuous
distribution \code{polygon} is used whereas for a discrete distribution \code{points} and \code{lines}
is used. Therefore the further graphical parameters must be set accordingly.
}
\details{
Note: for the distribution and quantile function of a discrete distribution an error is generated!
}
\examples{
d <- distrFunc("norm")
plot(d)
interval(d, -1.96, +1.96, col="red")
d <- distrFunc("geom", prob=0.7)
plot(d)
interval(d, 1, 5, col="red")
}
