% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{is.distrFunc}
\alias{is.distrFunc}
\title{is.distrFunc}
\usage{
is.distrFunc(x)
}
\arguments{
\item{x}{an \code{R} object to be tested}
}
\value{
returns if \code{x} is a distribution function object
}
\description{
The function \code{is.distrFunc} indicates whether \code{x} is a distribution function.
}
\examples{
d <- distrFunc("norm")
is.distrFunc(d)
is.distrFunc(pi)
}
