% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{is.widget}
\alias{is.widget}
\title{is.widget}
\usage{
is.widget(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is a widget or not.
}
\description{
Tests if object \code{x} is a widget
}
\examples{
w <- widgetHypergeometric('test')
is.widget(w)
is.widget("no widget")
}
