% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{loadPackages}
\usage{
loadPackages(libnames, ...)
}
\arguments{
\item{libnames}{character: names of libraries to load}
}
\value{
boolean: if a library could be successfully loaded
}
\description{
Tries to load a package with \code{require}. If it fails then installs the package via \code{install.packages} 
and tries to reload it again with \code{library}.
}
\examples{
loadPackages('shiny', 'shinydashboard')
}
