% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyTable.R
\name{median.frequencyTable}
\alias{median.frequencyTable}
\title{median.frequencyTable}
\usage{
\method{median}{frequencyTable}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{frequency table object}

\item{na.rm}{logical: unused}

\item{...}{further parameters}
}
\value{
median
}
\description{
Computes the median based on the frequency table.
}
\examples{
x  <- runif(100)
median(x)
ft <- frequencyTable(x)
median(ft)
}
