% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{mergeListsByName}
\alias{mergeListsByName}
\title{mergeListsByName}
\usage{
mergeListsByName(...)
}
\arguments{
\item{...}{list: a list of named lists}
}
\value{
a merged list
}
\description{
Merges two or more lists by element names. If two lists contain an element with the same name then the 
element from the last list will become the element of the resulting list.
}
\examples{
default.parameters <- list(a=1, b=1)
use.parameters <- list(a=3, c=1)
mergeListsByName(default.parameters, use.parameters)
}
