% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{mmstatPar}
\alias{mmstatPar}
\title{mmstatPar}
\usage{
mmstatPar(...)
}
\arguments{
\item{...}{list: list of named parameter to set a parameter or characters to question the current
value of parameter in mmstat}
}
\value{
a named list with the old parameter values
}
\description{
Sets and queries parameter values by mmstat.
}
\details{
The following parameters are available for setting and getting
\itemize{
\item{\code{col}}{ a vector of eight light and dark colors which are generated with the R Color brewer}
\item{\code{col.sample}}{ color used in mmstat to draw sample data/parameters}
\item{\code{col.sample.robust}}{ color used in mmstat to draw robust sample parameters}
\item{\code{col.population}}{ color used in mmstat to draw population data/parameters}
\item{\code{col.population.robust}}{ color used in mmstat to draw robust population parameters}
\item{\code{alpha}}{ the default set for possible significance levels}
}
}
\examples{
mmstatPar('col.sample')
mmstatPar(col.sample = "blue", col.population = "red")
}
