% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{permutation}
\alias{permutation}
\title{permutation}
\usage{
permutation(n, ...)
}
\arguments{
\item{n}{numeric: number of elements to select from}

\item{...}{numeric: number of repetitions copies}
}
\value{
numeric
}
\description{
Permutations are arrangements of objects (with or without repetition), order does matter.
A permutation with repetition is an arrangement of objects, where some objects are repeated a 
prescribed number of times.
}
\details{
\itemize{
\item{without repetition}{\eqn{P(n)=n!}}
\item{with repetition}{\eqn{P(n,k_1,k_2,...)=\frac{n!}{k_1!k_2!...}}}
}
}
\note{
If \eqn{k_1+k_2+... < n} then it assumed that for the rest the number of copies is always one.
}
\examples{
permutation(10)
permutation(10, 5, 3) # equivalent to permutation(10, 5, 3, 1, 1)
permutation(10, 5, 3, 2) 
}
