% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{plot.distrFunc}
\alias{plot.distrFunc}
\title{plot.distrFunc}
\usage{
\method{plot}{distrFunc}(type, add = F, xlim = NULL, ylim = NULL,
  cex.lettering = 1, n = 501, ...)
}
\arguments{
\item{type}{distribution function object}

\item{add}{logical: add to current plot (default: add=F)}

\item{xlim}{range: plotting area (default: NULL)}

\item{ylim}{range: plotting area (default: NULL)}

\item{cex.lettering}{numeric: magnification size for fonts used (default: 1)}

\item{n}{numeric: number of points used to draw the curve (default: 201)}

\item{...}{further drawing parameter, e.g color etc.}
}
\description{
Plots a distribution function. For a discrete distribution a needle plot (mass) or an stepwise 
function (distributiion, quantile) is used. For a continuous distribution simple lines 
are used. Since internally \code{points} and \code{lines} is used, the further drawing parameters
relate to that.
}
\examples{
## Standard normal distribution (continuous)
# density
d <- distrFunc("norm")
plot(d)
# distribution
d <- distrFunc("norm", func="p")
plot(d)
# quantile
d <- distrFunc("norm", func="q")
plot(d)
## Geoimetric distribution (continuous)
# density
d <- distrFunc("geom", prob=0.5)
plot(d)
# distribution
d <- distrFunc("geom", prob=0.5, func="p")
plot(d)
# quantile
d <- distrFunc("geom", prob=0.5, func="q")
plot(d)
}
