% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyTable.R
\name{plot.frequencyTable}
\alias{plot.frequencyTable}
\title{plot.frequencyTable}
\usage{
\method{plot}{frequencyTable}(x, ...)
}
\arguments{
\item{x}{frequency table object}

\item{...}{further parameters used \code{\link[graphics]{hist}}}
}
\value{
plot
}
\description{
Plots a frequency table as histogram.
}
\examples{
ft <- frequenyTable(runif(100))
plot(ft)
}
