% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetLanguage.R
\name{poText}
\alias{poText}
\title{Text translation: getText}
\usage{
poText(obj, env = NULL)
}
\arguments{
\item{obj}{R object:}

\item{env}{widget: language widget}
}
\value{
a R object witj translated texts
}
\description{
Each text element of the R object \code{obj} is translated in the current target language. 
If the translation of a text element is not possible then the original text is delivered.
}
\details{
For the translations must be PO files provided. Since \code{\link{getText}} is used 
internally the same holds as in \code{\link{getText}} given. Currently are only tables as 
objects supported, all other objects will deliver a warning.
}
\examples{
lang <- widgetLanguage ('lang')
tab  <- table(round(rnorm(100)), round(rnorm(100)))
dimnames(tab) <- c("GERMAN", "ENGLISH")
poText(tab, lang)
}
