% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{print.distrFunc}
\alias{print.distrFunc}
\title{print.distrFunc}
\usage{
\method{print}{distrFunc}(x, ...)
}
\arguments{
\item{x}{distribution function object}

\item{...}{unused}
}
\value{
Returns invisibly \code{x}.
}
\description{
Prints information about a distribution function.
}
\examples{
dn <- distrFunc("norm")
dn
dg <- distrFunc("geom", prob=0.7)
dg
}
