% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{printVar}
\alias{printVar}
\title{Prints the contents of an R object}
\usage{
printVar(v)
}
\arguments{
\item{v}{R object}
}
\value{
Shows the contents of \code{v} using \code{str}
}
\description{
Prints the contents of an R object
}
\examples{
x <- 1:10
printVar(x)
}
