% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderPanel.R
\name{renderPanel}
\alias{renderPanel}
\title{Render a conditional panel}
\usage{
renderPanel(inputId, ..., session = NULL)
}
\arguments{
\item{inputId}{input element for the condition}

\item{...}{parameter sets for the widget to render}

\item{session}{session parameter of the server function}

\item{lang}{language widget (or NULL)}
}
\value{
HTML code which can be used instead of \code{renderUI}
}
\description{
Render a conditional panel
}
\examples{
\dontrun{
  demo(mmstatDiscreteDistributions)
}
}
\seealso{
\code{\link{conditionalPanel}}
}
