% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderWidget.R
\name{renderWidget}
\alias{renderWidget}
\title{Renders a widget}
\usage{
renderWidget(env, session = NULL, elem = NULL)
}
\arguments{
\item{env}{widget}

\item{session}{session parameter of the server function}

\item{elem}{element of widget to render, if NULL the whole widget will be rendered}
}
\value{
HTML code which can be used instead of \code{renderUI}
}
\description{
Renders the complete widget or a element of it. 
 
If the parameter \code{lang} is not NULL then it should contain the number of the 
selected language, usually it will be the \code{input[[getInputs(language widget)]]}.
If the parameter \code{session} is not NULL and the \code{value} of a slider element
is set to "session" then the starting value is taken from the URL.
}
\details{
E.g. let \code{l <- widgetLanguage('lang')} then the input element will be called 
\code{lang.language}. Calling the Shiny app with \code{http://...?lang.language=2} will
result that app start with second language as default rather then with the first.
}
\note{
A starting value of a widget element by URL can be only set at the beginning of 
session. Reloading the page will NOT start a new shiny session.
}
\examples{
w <- widgetFontSize('widget')
renderWidget(w)

\dontrun{
  # Open in browser and add "?widget.size=11" to the URL to set the starting
  # value to 11
  shinyDemo('testBinomialSession')
}
}
