% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{stopif}
\alias{stopif}
\title{Stops execution}
\usage{
stopif(cond, fmt, ...)
}
\arguments{
\item{cond}{condition, execution is stopped if any entry is \code{TRUE}}

\item{fmt}{a character vector of format strings for \code{sprintf}}

\item{...}{further values to be passed into \code{fmt}. Only logical, integer, real and character vectors are supported.}
}
\description{
Stops the execution of a Shiny app, if at least \code{cond} contains one \code{TRUE}. The function uses \code{validate} and \code{need} from the library \code{shiny}.
}
\examples{
\dontrun{
  # Produces an error
  x <- -1
  stopif(x<=0, "x is smaller than \%i", 0)
}
}
