% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{str.widget}
\alias{str.widget}
\title{str}
\usage{
\method{str}{widget}(object, ...)
}
\arguments{
\item{object}{widget}

\item{...}{further parameters (currently ignored)}
}
\description{
Compactly display the internal structure of a widget
}
\examples{
w <- widgetHypergeometric('test')
str(w)
}
