% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{summary.widget}
\alias{summary.widget}
\title{summary}
\usage{
\method{summary}{widget}(w)
}
\arguments{
\item{w}{widget}
}
\value{
Informations about the widget
Widget name    : widget name
UI element(s)  : list of Shiny UI elements
Observe values : number of values which are observed
Observe changes: number of values which changes are observed
Observer       : does the widget has an observer
Language       : does the widget contain a language element (for internationalization)
Internal data  : does the widget contain internal data
}
\description{
Summarizes the widget contents
}
\examples{
w <- widgetHypergeometric('test')
summary(w)
}
