% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2d.R
\name{table2d}
\alias{table2d}
\title{table2d}
\usage{
table2d(tab, horizontal = FALSE, header = NULL)
}
\arguments{
\item{tab}{table: one dimensional table to convert}

\item{horizontal}{logical: horizontal or vertical table (default: \code{FALSE})}

\item{header}{character: header for the single column or row}
}
\value{
a two dimensional table
}
\description{
Converts a one dimensional table to a two dimensional table.
}
\examples{
tab <- table(round(rnorm(100)))
tab
table2d(tab)
str(tab)
str(table2d(tab, horizontal=TRUE))
}
