% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2html.R
\name{table2html}
\alias{table2html}
\title{table2html}
\usage{
table2html(tab, colsums = NULL, rowsums = NULL, bg = c("white",
  "lightGray"), border = "grey", style = "width:95\%;margin:10px;", ...)
}
\arguments{
\item{tab}{table: table to show}

\item{colsums}{character: If non-NULL the column sums will be computed and the contents of \code{colsums} appears as header (default: \code{NULL})}

\item{rowsums}{character: If non-NULL the rows sums will be computed and the contents of \code{rowsums} appears as header (default: \code{NULL})}

\item{bg}{colors: background color for table rows, will be recycled (default: \code{c("white", "lightGray")})}

\item{border}{color: background color for header cells (default: \code{"grey"})}

\item{style}{character: style information for the table (default: \code{"width:95\%;margin:10px;"})}

\item{...}{further parameters used for the \code{\link{format}} command for formatting table entries}
}
\value{
character: the HTML code for the table
}
\description{
Creates an HTML view on a two dimensional table.
}
\details{
In case of a non-numeric table \code{colsums} and \code{rowsums} is ignored.
}
\examples{
tab <- HairEyeColor[,,'Female']
htmltab <- table2html(tab, colsums='', rowsums='')
sink('HairEye.html')
cat(htmltab)
sink()
}
