% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{variation}
\alias{variation}
\title{variation}
\usage{
variation(n, k, rep = F)
}
\arguments{
\item{n}{numeric: number of elements to select from}

\item{k}{numeric: number of elements selected}

\item{rep}{logical: with or without repetition (default: FALSE)}
}
\value{
numeric
}
\description{
Variations are arrangements of selections of objects, where the order of the selected objects matters.
\code{variation} computes the number of \code{k}-element variations of \code{n} objects either with repetition 
or not.
}
\details{
\itemize{
\item{without repetition}{\eqn{V(n,k)=\frac{n!}{(n-k)!}}}
\item{with repetition}{\eqn{V(n,k)=n^k}}
}
}
\examples{
variation(10, 3)
variation(10, 3, TRUE)
}
