% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrFunc.R
\name{vline}
\alias{vline}
\title{vline}
\usage{
vline(dist, x, cont = 1, ...)
}
\arguments{
\item{dist}{distribution object}

\item{x}{numeric: x-coordinates}

\item{cont}{numeric: continuation (default: 1)}

\item{...}{further graphical parameters, e.g. line color etc.}
}
\value{
Returns invisibly a list with x-y-coordinates of the intersection of x coordinates 
with the distribution function.
}
\description{
Draws one (or more) vertical line(s) in a plot. The line is drawn from the bottom until it 
intersects the distribution function and either does not continut (\code{cont=1}, default), 
continues to the left (\code{cont=2}), to the top (\code{cont=3}) or to the right (\code{cont=4}).
}
\examples{
d <- distrFunc("norm")
plot(d)
vline(d, -1.96, cont=2, col="blue")
vline(d, 0)
vline(d, +1.96, cont=4, col="blue")
}
