% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{widget}
\alias{widget}
\title{widget}
\usage{
widget(inputId, lang)
}
\arguments{
\item{inputId}{widget name}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
Creates a new widget. Basically a widget is an \code{\link{environment}} with several components
}
\details{
\describe{
\item{\code{ui$elem}}{shiny UI element with (fixed) name \code{elem}}
\item{\code{ui$elem$func}}{shiny UI widget, e.g. \code{\link{sliderInput}}, ...}
\item{\code{ui$elem$args}}{\code{\link{do.call}} argument list for shiny UI widget func}
\item{\code{change$elem}}{store current value of \code{ui$elem} (intended for buttons)}
\item{\code{value$elem}}{store current value of \code{ui$elem}}
\item{\code{selected}}{store current number of selected language}
\item{\code{widgetId}}{name of widget}
\item{\code{observe}}{expression which contains code for observe}
\item{\code{lang}}{language widget (necessary for translation)}
\item{\code{data}}{internal data the widget needs}
}
}
\examples{
# empty widget with no UI elements
w <- widget('test', NULL)
}
