% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetActionButton}
\alias{widgetActionButton}
\title{Encapsulates a Shiny \code{actionButton}}
\usage{
widgetActionButton(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{actionButton}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{actionButton}. If you do not need a language change on-the-fly then you should use \code{actionButton}.
}
\examples{
widgetActionButton(list(inputId="goButton", label='Go!'), lang=NULL)
}
\seealso{
\code{\link{actionButton}}
}
