% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetBinomial.R
\name{widgetBinomial}
\alias{widgetBinomial}
\title{Widget for a binomial distribution}
\usage{
widgetBinomial(inputId, size = list(), prob = list(), lang = NULL)
}
\arguments{
\item{inputId}{character: widget name}

\item{size}{list: \code{sliderInput} parameter for the number of draws (default: \code{list()}, see Details)}

\item{prob}{list: \code{sliderInput} parameter for the probability of success per draw (default: \code{list()}, see Details)}

\item{lang}{widget: language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for a binomial distribution
}
\examples{
\dontrun{
  shinyDemo('testBinomial')
  shinyDemo('testBinomialLang')
  shinyDemo('mmstatDiscreteDistribution')
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{dbinom}}
}
