% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetCheckbox}
\alias{widgetCheckbox}
\title{Encapsulates a Shiny checkboxInput}
\usage{
widgetCheckbox(param, lang)
}
\arguments{
\item{param}{list of parameters for checkboxInput}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny checkboxInput. If you do not need a language change on-the-fly then you should use checkboxInput.
}
\examples{
widgetCheckbox(list(inputId="outliers", label='Show outliers'), lang=NULL)
}
\seealso{
\code{\link{checkboxInput}}
}
