% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetCheckboxGroup}
\alias{widgetCheckboxGroup}
\title{Encapsulates a Shiny \code{checkboxGroupInput}}
\usage{
widgetCheckboxGroup(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{checkboxGroupInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{checkboxGroupInput}. If you do not need a language change on-the-fly then you should use \code{checkboxGroupInput}.
}
\examples{
widgetCheckboxGroup(list(inputId="variable", label='Variable', 
                         choices=list("Cylinders" = "cyl", 
                                      "Transmission" = "am", 
                                      "Gears" = "gear")), 
                    lang=NULL)
}
\seealso{
\code{\link{checkboxGroupInput}}
}
