% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetConfidenceLevel.R
\name{widgetConfidenceLevel}
\alias{widgetConfidenceLevel}
\title{Widget for a confidence level}
\usage{
widgetConfidenceLevel(inputId, level = list(), lang = NULL)
}
\arguments{
\item{inputId}{character: widget name}

\item{level}{list: \code{sliderInput} parameter for the confidence level  (default: \code{list()}, see Details)}

\item{lang}{widget: language widget}
}
\value{
a widget object (environment)
}
\description{
The widget allows to choose a confidence level on a percent basis. 
This means for a 95\% confidence level the result is 95 (and not 0.95).
}
\details{
For \code{level} a named list with the parameters for 
\code{\link{sliderInput}} can be given. If just a numeric value is given then the according 
UI element is not drawn, but this value delivered back by \code{getValues}.
}
\examples{
\dontrun{
  # Press ESC after finishing the app
  demo(testConfidenceLevel)
  demo(testConfidenceLevelLang)
}
}
\seealso{
\code{\link{sliderInput}}
}
