% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetDataset.R
\name{widgetDataset}
\alias{widgetDataset}
\title{Widget for data set and variable selection}
\usage{
widgetDataset(inputId, pattern = c("^[[:alnum:]]+.rds$",
  "^[[:alnum:]]+.RDS$"), pathes = c(".", find.package("mmstat")),
  duplicates.ok = FALSE, datasettype = is.data.frame, data = list(),
  lang = NULL)
}
\arguments{
\item{inputId}{character: widget name}

\item{pattern}{character: a \link{regular expression}. Only file names which match the regular
expression will be read (default: \code{c('^[[:alnum:]]+.rds$', '^[[:alnum:]]+.RDS$')})}

\item{pathes}{character: directories to read data files from (default: \code{c('.', find.package('mmstat'))})}

\item{duplicates.ok}{logical: can variables be in a selection be used more than once? (default: FALSE)}

\item{datasettype}{logical function: should data set used? (default: \code{is.data.frame})}

\item{data}{list: \code{selectInput} parameter for the data set(s) (default: \code{list()})}

\item{lang}{widget: language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Reads in all \code{*.rds} or \code{*.RDS} files from the first non-empty directory in \code{pathes}
and extracts the variables. Each file must contain one data frame only.
The widget must be used together with \code{widgetVariables}.
}
\examples{
\dontrun{
  shinyDemo('testDataset1') # one variable
  shinyDemo('testDataset2') # two variables
  shinyDemo('testDataset3') # two variables: one numeric, one factor
  shinyDemo('testDataset3Lang')
}
}
\seealso{
\code{\link{selectInput}}, \code{\link{sliderInput}}, \code{\link{widgetVariables}}
}
