% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetDate}
\alias{widgetDate}
\title{Encapsulates a Shiny \code{dateInput}}
\usage{
widgetDate(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{dateInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{dateInput}. If you do not need a language change on-the-fly then you should use \code{dateInput}.
}
\examples{
widgetDate(list(inputId='date', label='Date:', value = '2012-02-29'), lang=NULL)
}
\seealso{
\code{\link{dateInput}}
}
