% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetDrawSample.R
\name{widgetDrawSample}
\alias{widgetDrawSample}
\title{Widget for drawing a samples(s) from a population}
\usage{
widgetDrawSample(inputId, size = list(), sample = list(), reset = list(),
  speed = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{size}{list: \code{sliderInput} parameter for the sample size (or NULL)}

\item{sample}{\code{actionButton} parameter for drawing one sample}

\item{reset}{\code{actionButton} parameter to reset the sample drawing (or NULL)}

\item{speed}{\code{sliderInput} parameter for repeatedly drawing samples (or NULL)}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
If the widget elements \code{size}, \code{reset} or \code{speed} are set to NULL then 
they will not display or used.
}
\examples{
\dontrun{
  shinyDemo('testDrawSample')
  shinyDemo('testDrawSampleLang')
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{actionButton}}
}
