% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetFile}
\alias{widgetFile}
\title{Encapsulates a Shiny \code{fileInput}}
\usage{
widgetFile(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{fileInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{fileInput}. If you do not need a language change on-the-fly then you should use \code{fileInput}.
}
\examples{
widgetFile(list(inputId='file', label='File:'), lang=NULL)
}
\seealso{
\code{\link{fileInput}}
}
