% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetFontSize.R
\name{widgetFontSize}
\alias{widgetFontSize}
\title{Widget for font size choice}
\usage{
widgetFontSize(inputId, fontsize = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{fontsize}{\code{sliderInput} parameters for font size}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for font size choice
}
\examples{
\dontrun{
  # Press ESC after finishing the demo
  demo(testFontSize)
  demo(testFontSizeLang)
}
}
\seealso{
\code{\link{sliderInput}}
}
