% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelpButton.R
\name{widgetHelpButton}
\alias{widgetHelpButton}
\title{Widget for a Help Button}
\usage{
widgetHelpButton(inputId, url, lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{url}{the html address regading the help page to display}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for a Help Button
}
\examples{
\dontrun{
  # Press ESC after finishing the app
}
}
\seealso{
\code{\link{actionButton}}
}
