% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHypergeometric.R
\name{widgetHypergeometric}
\alias{widgetHypergeometric}
\title{Widget for a hypergeometric distribution}
\usage{
widgetHypergeometric(inputId, N = list(), M = list(), n = list(),
  lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{N}{\code{sliderInput} parameter for the population size}

\item{M}{\code{sliderInput} parameter for the number of success states in the population}

\item{n}{\code{sliderInput} parameter for the number of draws}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
This widget needs an \code{\link{widgetObserve}} to update \code{M} and \code{n} if \code{N} changes.
}
\details{
Note that the widget does not follow the parameter definition of \code{dhyper}. 
The parameter \code{N} from \code{widgetHypergeometric} is equal to \code{m+n} from \code{dhyper}.
}
\examples{
\dontrun{
  # Press ESC after finishing the demo
  demo(testHypergeometric)
  demo(testHypergeometricLang)
  demo(mmstatDiscreteDistribution)
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{dhyper}}
}
