% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetIntensity.R
\name{widgetIntensity}
\alias{widgetIntensity}
\title{Widget for a Poisson or exponential distribution}
\usage{
widgetIntensity(inputId, lambda = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{lambda}{\code{sliderInput} parameter for the intensity parameter}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for a Poisson or exponential distribution
}
\examples{
\dontrun{
  # Press ESC after finishing the demo
  demo(testIntensity)
  demo(testIntensityLang)
  demo(mmstatDiscreteDistribution)    # see Poisson distribution
  demo(mmstatContinuousDistribution)  # see Exponential distribution
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{dpois}}, \code{\link{dexp}}
}
