% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetLanguage.R
\name{widgetLanguage}
\alias{widgetLanguage}
\title{Widget for language choice}
\usage{
widgetLanguage(inputId, lang = list(), path = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{lang}{\code{selectInput} parameter for selecting a language}

\item{path}{directory where the PO files are found. The default location is the directory where
the R the \code{shinyWidgets} package stores.}
}
\value{
a widget object (environment)
}
\description{
For the translations must be PO files provided. Beside creating the widget also the PO file 
are read and stored internally.
}
\details{
To find the location of the PO files use 
\code{system.file(package='shinyWidgets')}
}
\examples{
\dontrun{
  # Press ESC to finish the app(s)
  demo(testLanguage)
}
}
\seealso{
\code{\link{selectInput}}
}
