% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetNormal.R
\name{widgetNormal}
\alias{widgetNormal}
\title{Widget for a normal distribution}
\usage{
widgetNormal(inputId, mean = list(), sd = list(), lang = NULL)
}
\arguments{
\item{inputId}{character: widget name}

\item{mean}{\code{sliderInput} parameter for the mean of the normal distribution (default: \code{list()}, see Details)}

\item{sd}{\code{sliderInput} parameter for the standard deviation of the normal distribution (default: \code{list()}, see Details)}

\item{lang}{widget: language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for a normal distribution
}
\details{
For \code{mean} and \code{sd} a named list with the parameters for 
\code{\link{sliderInput}} can be given. If just a numeric value is given then the according 
UI element is not drawn, but this value delivered back by \code{getValues}.
}
\examples{
\dontrun{
  # Press ESC after finishing the demo
  demo(testNormal)
  demo(testNormalConst)
  demo(testNormalLang)
  demo(testNormalInputId)
  demo(mmstatContinuousDistribution)
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{dnorm}}
}
