% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetObserve.R
\name{widgetObserve}
\alias{widgetObserve}
\title{Create a reactive observer}
\usage{
widgetObserve(env, input, session)
}
\arguments{
\item{env}{widget object}

\item{input}{input object: \code{input} parameter from the server function of the app}

\item{session}{session object: \code{session} parameter from the server function of the app}
}
\value{
a observer for the Shiny app
}
\description{
Creates an observer from the given widget. The observer handles internal relationships between the elements of a widget.
}
