% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetPCDF.R
\name{widgetPCDF}
\alias{widgetPCDF}
\title{Widget for selecting between functions}
\usage{
widgetPCDF(inputId, pcdf = list(), refit = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{pcdf}{\code{radioButtons} parameter for the function selection}

\item{refit}{\code{actionButton} parameter to reset the axes}

\item{lang}{language widget}
}
\value{
a widget object (environment)
}
\description{
You can select between a probability mass function/probability density function.
A button allows to refit the axes to current display function.
}
\examples{
\dontrun{
  # Press ESC after finishing the app
  shinyDemo('testPCDF')
  shinyDemo('testPCDFLang')
}
}
\seealso{
\code{\link{radioButtons}}, \code{\link{actionButton}}
}
