% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetPercentage.R
\name{widgetPercentage}
\alias{widgetPercentage}
\title{Widget for a percentage value}
\usage{
widgetPercentage(inputId, p = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{p}{\code{sliderInput} parameter for the percentage}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Widget for a percentage value
}
\examples{
\dontrun{
  shinyDemo('testPercentage')
  shinyDemo('testPercentageLang')
}
}
