% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetRadioButtons}
\alias{widgetRadioButtons}
\title{Encapsulates a Shiny \code{radioButtons}}
\usage{
widgetRadioButtons(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{radioButtons}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{radioButtons}. If you do not need a language change on-the-fly then you should use \code{radioButtons}.
}
\examples{
widgetRadioButtons(list(inputId='dist', label='Distribution type:', 
                        choices=list("Normal" = "norm",
                                     "Uniform" = "unif", 
                                     "Log-normal" = "lnorm",
                                     "Exponential" = "exp")), 
                   lang=NULL)
}
\seealso{
\code{\link{radioButtons}}
}
