% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetSelect}
\alias{widgetSelect}
\title{Encapsulates a Shiny \code{selectInput}}
\usage{
widgetSelect(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{selectInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{selectInput}. If you do not need a language change on-the-fly then you should use \code{selectInput}.
}
\examples{
widgetSelect(list(inputId="variable", label="Variable:",
                  choices=list("Cylinders" = "cyl",
                               "Transmission" = "am",
                               "Gears" = "gear")),
             lang=NULL)
}
\seealso{
\code{\link{selectInput}}
}
