% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetShiny}
\alias{widgetShiny}
\title{Encapsulates a Shiny input element}
\usage{
widgetShiny(input, param, lang, tooltip = NULL)
}
\arguments{
\item{input}{input element: either a function or function name}

\item{param}{list: parameters for the input element}

\item{lang}{language widget (or NULL)}

\item{tooltip}{character or list: text of tooltip or parameter list for \code{\link[shinyBS]{bsTooltip}}}
}
\value{
a widget object (environment)
}
\description{
This the basic function to encapsulate all shiny input elements. If you do not need a language change on-the-fly then you should use shiny input element.
traceback()
}
\examples{
library('shiny')
widgetShiny(sliderInput, list(inputId='slider', label='My slider', min=0, max=10), lang=NULL)
}
