% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetSlider}
\alias{widgetSlider}
\title{Encapsulates a Shiny \code{sliderInput}}
\usage{
widgetSlider(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{sliderInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{sliderInput}. If you do not need a language change on-the-fly then you should use \code{sliderInput}.
}
\examples{
widgetSlider(list(inputId='slider', label='My slider', min=0, max=10), lang=NULL)
}
\seealso{
\code{\link{sliderInput}}
}
