% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetTest.R
\name{widgetTest}
\alias{widgetTest}
\title{Widget for a one parameter test}
\usage{
widgetTest(inputId, alternative = list(), parameter = list(),
  level = list(), lang = NULL)
}
\arguments{
\item{inputId}{widget name}

\item{alternative}{\code{selectInput} parameter for the alternative: two sided, less or greater (or NULL)}

\item{parameter}{\code{sliderInput} parameter for the hypothetical parameter}

\item{level}{\code{sliderInput} parameter for the significance level (or NULL)}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
If \code{alternative} or \code{level} ist set to NULL then the corresponding widget element is not displayed.
}
\examples{
\dontrun{
  # Press ESC after finishing the app
  demo(testTest)
  demo(testTestLang)
}
}
\seealso{
\code{\link{selectInput}}, \code{\link{sliderInput}}
}
