% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetShiny.R
\name{widgetText}
\alias{widgetText}
\title{Encapsulates a Shiny \code{textInput}}
\usage{
widgetText(param, lang)
}
\arguments{
\item{param}{list of parameters for \code{textInput}}

\item{lang}{language widget (or NULL)}
}
\value{
a widget object (environment)
}
\description{
Encapsulates a Shiny \code{textInput}. If you do not need a language change on-the-fly then you should use \code{textInput}.
}
\examples{
widgetText(list(inputId='caption', label='Caption:', value="Data summary"), lang=NULL)
}
\seealso{
\code{\link{textInput}}
}
