% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetHelper.R
\name{widgetValueChanged}
\alias{widgetValueChanged}
\title{Is a widget value changed}
\usage{
widgetValueChanged(env, input)
}
\arguments{
\item{env}{a widget object}

\item{input}{current values of all of the widgets in your app. Is the first parameter of the \code{server} function.}
}
\value{
named list with the logical values
}
\description{
If you use a button then the corresponding element in \code{input} delivers the number of times
the button was pressed. \code{widgetValueChanged} returns a list with information if the 
internally  stored value is different from the corresponding element in \code{input} and 
stores the new value.
}
\examples{
\dontrun{
  # Press ESC after finishing the app
  shinyDemo('testDrawSample')
  shinyDemo('testDrawSampleLang')
}
}
\seealso{
\code{\link{widgetActionButton}}
}
