% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetDataset.R
\name{widgetVariables}
\alias{widgetVariables}
\title{Widget to select variables of a dataset}
\usage{
widgetVariables(env, ...)
}
\arguments{
\item{env}{widget: data set widget}

\item{...}{lists: parameter lists for \code{selectInput} to select one variable}
}
\description{
This widget needs an \code{\link{widgetObserve}} to update data set changes to the variable selectors.
}
\details{
Usually all variables of data frame are offered. If the parameter list for \code{selectInput}s
contain a parameter \code{vartype=func} then only the variables can be selected for which the
\code{func} delivers \code{TRUE}. An example would be \code{vartype=is.factor} such that the
variable selector just offers the categorical variables of the data frame.
The widget must be used together with \code{widgetDataset}.
}
\note{
The function will produce a Shiny error if not all data sets read contain at least one
appropiate variable.
}
\examples{
\dontrun{
  # Press ESC after finishing the app
  demo(testDataset1) # one variable
  demo(testDataset2) # two variables
  demo(testDataset3) # two variables: one numeric, one factor
  demo(testDataset3Lang)
}
}
\seealso{
\code{\link{selectInput}}, \code{\link{sliderInput}}, \code{\link{widgetDataset}}
}
