% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTable.R
\name{association}
\alias{association}
\title{coefficients}
\usage{
association(ft, ...)
}
\arguments{
\item{ft}{contingency table object}

\item{...}{list: coefficients in the table}
}
\value{
a two dimensional table
}
\description{
Extracts all requested association coefficients as two dimensional table. Possible names are:
}
\details{
\describe{
\item{\code{chisq}}{Chi square value}
\item{\code{C}}{Contingency value}
\item{\code{CC}}{Corrected c ontingency value}
\item{\code{V}}{Cramers V or phi}
\item{\code{lambda}}{All three Goodman and Kruskal lambdas}
\item{\code{lambda.cr}}{Goodman and Kruskal lambda, column dependent}
\item{\code{lambda.rc}}{Goodman and Kruskal lambda, row dependent}
\item{\code{lambda.sym}}{Goodman and Kruskal lambda, symmetric}
\item{\code{gkt}}{Both Goodman and Kruskal taus}
\item{\code{gkt.cr}}{Goodman and Kruskal tau, column dependent}
\item{\code{gkt.rc}}{Goodman and Kruskal tau, row dependent}
\item{\code{theil}}{All three uncertainty coefficients/Theil's U}
\item{\code{theil.cr}}{Uncertainty coefficient/Theil's U, column dependent}
\item{\code{theil.rc}}{Uncertainty coefficient/Theil's U, row dependent}
\item{\code{theil.sym}}{Uncertainty coefficient/Theil's U, symmetric}
}
}
\examples{
ct <- contingencyTable(HairEyeColor[,,"Female"])
coefficients(ct, lambda=TRUE)
}
