% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getValues}
\alias{getValues}
\title{Input values for a widget}
\usage{
getValues(env, input)
}
\arguments{
\item{env}{widget}

\item{input}{current values of all of the widgets in your app. Is the first parameter of the \code{server} function.}
}
\value{
named list with the values
}
\description{
Input values for a widget
}
\examples{
w <- widgetFontSize('widget')
getValues(w, NULL)
}
