% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetClass.R
\name{print.widget}
\alias{print.widget}
\title{print}
\usage{
\method{print}{widget}(w)
}
\arguments{
\item{w}{widget}
}
\value{
the widget
}
\description{
Prints the used Shiny UI elements and its parameter. Returns it invisibly (via \code{\link{invisible}(w)}).
}
\examples{
w <- widgetHypergeometric('test')
print(w)
}
