% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDemo.R
\name{shinyDemo}
\alias{shinyDemo}
\title{shinyDemo}
\usage{
shinyDemo(example, launch.browser = FALSE)
}
\arguments{
\item{example}{character: name of the example}

\item{launch.browser}{logical:
If true, the system's default web browser will be launched automatically after the app is started (default: \code{FALSE})}
}
\description{
Runs a shiny app similar to \code{demo}.
}
\examples{
\dontrun{
  # to view all possible apps
  shinyDemo()
  # to run one app 
  shinyDemo('mmstatContinuousDistribution')
}
}
\references{
Based on \href{http://deanattali.com/2015/04/21/r-package-shiny-app/}{Supplementing your R package with a Shiny app} by Dean Attali.
}
