% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDemo.R
\name{shinyDemo}
\alias{shinyDemo}
\title{shinyDemo}
\usage{
shinyDemo(example)
}
\arguments{
\item{example}{character: name of the example}
}
\description{
Runs a shiny app similar to \code{demo}.
}
\examples{
shinyDemo('mmstatContinuousDistribution')
}
\references{
Based on \href{http://deanattali.com/2015/04/21/r-package-shiny-app/}{Supplementing your R package with a Shiny app} by Dean Attali.
}
