% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetBinomial.R
\name{widgetBinomial}
\alias{widgetBinomial}
\title{Widget for a binomial distribution}
\usage{
widgetBinomial(inputId, size = list(), prob = list(), lang = NULL)
}
\arguments{
\item{inputId}{character: widget name}

\item{size}{list: \code{sliderInput} parameter for the number of draws (default: \code{list()}, see Details)}

\item{prob}{list: \code{sliderInput} parameter for the probability of success per draw (default: \code{list()}, see Details)}

\item{lang}{widget: language widget}
}
\value{
a widget object (environment)
}
\description{
Widget for a binomial distribution
}
\details{
Possible values for  \code{size} and \code{prob} are
\describe{
  \item{\strong{a named list}}{for the parameters for the according Shiny UI element, e.g. \code{\link{sliderInput}}}
  \item{\strong{\code{NULL}}}{the Shiny UI element is not drawn and does \emph{not} return any value in \code{getValues}}
  \item{\strong{any other value}}{the Shiny UI element is not drawn and does return the given value in \code{getValues}}
}
}
\examples{
\dontrun{
  # Press ESC after finishing the demo
  demo(testBinomial)
  demo(testBinomialLang)
  demo(mmstatDiscreteDistribution)
}
}
\seealso{
\code{\link{sliderInput}}, \code{\link{dbinom}}
}
